# feat: xxhash checksum blocks