--- a/Cargo.toml
+++ b/Cargo.toml
@@ -14,5 +14,4 @@ sbbf-rs-safe = "0.3.2"
 serde = {version = "1.0.219", features = ["derive"]}
 serde_json = "1.0.140"
 siphasher = "1.0.1"
-
-
+xxhash-rust = {version = "0.8.15", features = ["xxh3"]}