Issue: https://github.com/mikerhodes/toykv/issues/14

Add a checksum to blocks and validate on reading. Uses xxhash,
xxh3_64. The xxh3 family is stated to have better performance and
broadly seems recommended. (per https://xxhash.com/).

There is still a performance hit, however.

before:

Running write() took 952ms.
Running read() 100000 times took 6718ms (0.067ms per read).
Running spawned read() 400000 times took 7885ms (0.019ms per read).

Running scan() 998 times took 3746ms (3.753ms per scan).

after:

Running write() took 901ms.
Running read() 100000 times took 6524ms (0.065ms per read).
Running spawned read() 400000 times took 8886ms (0.022ms per read).

Running scan() 998 times took 5694ms (5.706ms per scan).

See afbf446e2cd0455099b483b7efe3d15e25dfa67e commit (just before this
one) where we optimised the block decoding. This commit returns
the performance to where we were _before_ that optimisation pretty
much exactly. So overall, it's not too bad.